Optimize Fold<Vec<T>>
